Revert "Clean"
This reverts commit c4f6789e147859c4a7ec15abb0ccaf20f9222ec9.

Revert "more freq data for euc-kr"

This reverts commit ca9734da785a1e07fad8e692843d24925bb3e39c.

Revert "Clean"

This reverts commit 177707c18df1b9b317fb22da8354f9fb0de02fea.

Revert "Clean"

This reverts commit 45943b4860f9518e0768c50b334b23b866a8f843.

Revert "Clean"

This reverts commit 56feb2fdd60bda8889637702c66152ab1bc1580d.

Revert "added cp949"

This reverts commit 799c8c61f40014528e2c47683ca6e02dbf9a6fd8.

Revert "added mbcssm CP949"

This reverts commit 8ba37c3b0a6c8b65a11aa4270403996c5b9a4ac1.

Revert "Clean"

This reverts commit 7e510679c2f12ccaec40b537b141ee8e6b7ac3d8.

+++ src/nsUTF8Prober.cpp
@@ -102,7 +102,7 @@ float nsUTF8Prober::GetConfidence(void)
   if (mNumOfMBChar < 6 && mbCharRatio <= 0.6f)
   {
     for (PRUint32 i = 0; i < mNumOfMBChar; i++)
-      unlike *= std::powf(ONE_CHAR_PROB, (float)mNumOfMBChar);
+      unlike *= std::pow(ONE_CHAR_PROB, mNumOfMBChar);
     return (float)1.0 - unlike;
   }
   else
